Adjust air supply check and add Character removal mechanism
The air supply comparison value in the onStopFloatingToSurfaceOfFluid method in Zane.java was changed from 40 to 60. Furthermore, a "removeWhenFarAway" method was added to the Character.java file. This method checks for the presence of entities of the same class within a large bounding box and removes the current entity if there are others present.